Crate assembly

This crate is a collection of parsers and data types To enable reading of data from LEGO Universe game files.

Re-exports

pub use assembly_core as core;

pub use assembly_data::fdb;

pub use assembly_data::xml;

Modules

luz

The zone/world (*.luz) file format

lvl

The level (*.lvl) file format

pk

The pack (*.pk) files

pki

The pack index (*.pki) files

sd0

The segmented (*.sd0) compression format