Module assembly::sd0

The segmented (*.sd0) compression format

This format is used to deflate (zlib) the data served from the server to the client, and to use less space in the pack archives.

Serialization

[L:5] 's' 'd' '0' 0x01 0xff
repeated:
    [u32] length V (of the following chunk)
    [L:V] zlib stream (deflate with zlib header)

Modules

fs

Compression for use in a script

index

Segmented Index Files (si0)

read

std::io::Read adapters for *.sd0 reading

write

std::io::Write adapters for *.sd0 writing

Structs

Compression

When compressing data, the compression level can be specified by a value in this struct.

Constants

MAGIC

The magic bytes for the sd0 format

Functions

decode

Decode a byte slice into a vector

encode

Encode a byte slice into a vector